+++ dis-infra-core/tests/runtime.rs
@@ -1,5 +1,3 @@
-use dis_infra_core::runtime::default_runtime;
-
 #[test]
 fn runtime_with_config() {
     // let runtime = default_runtime().unwrap();
@@ -27,7 +25,7 @@ fn runtime_with_config() {
 
 #[test]
 fn runtime_with_code() {
-    let runtime = default_runtime().unwrap();
-
-    runtime.run(async {});
+    // let runtime = default_runtime().unwrap();
+    //
+    // runtime.run(async {});
 }