+++ dis-rs/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Function `crate::supported_protocol_versions` that returns a `Vec<ProtocolVersions>` supported by the crate.
+- Top level `Pdu` struct is now `Clone`.
+
 ### Changed
 
 ### Deprecated